+++ ThirdPartyAdapters/dtexchange/dtexchange/src/test/kotlin/com/google/ads/mediation/fyber/FyberMediationAdapterTest.kt
@@ -17,22 +17,34 @@ package com.google.ads.mediation.fyber
 import android.app.Activity
 import androidx.core.os.bundleOf
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.fyber.inneractive.sdk.external.BidTokenProvider
 import com.fyber.inneractive.sdk.external.InneractiveAdManager
+import com.fyber.inneractive.sdk.external.InneractiveAdSpot
+import com.fyber.inneractive.sdk.external.InneractiveAdSpotManager
 import com.fyber.inneractive.sdk.external.OnFyberMarketplaceInitializedListener
 import com.fyber.inneractive.sdk.external.OnFyberMarketplaceInitializedListener.FyberInitStatus
 import com.google.ads.mediation.adaptertestkit.AdErrorMatcher
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants
+import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_BID_RESPONSE
 import com.google.ads.mediation.adaptertestkit.assertGetSdkVersion
 import com.google.ads.mediation.adaptertestkit.assertGetVersionInfo
+import com.google.ads.mediation.adaptertestkit.createMediationBannerAdConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationConfiguration
+import com.google.ads.mediation.adaptertestkit.createMediationInterstitialAdConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationRewardedAdConfiguration
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.AdFormat
 import com.google.android.gms.ads.MobileAds
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback
+import com.google.android.gms.ads.mediation.MediationBannerAd
+import com.google.android.gms.ads.mediation.MediationBannerAdCallback
+import com.google.android.gms.ads.mediation.MediationInterstitialAd
+import com.google.android.gms.ads.mediation.MediationInterstitialAdCallback
 import com.google.android.gms.ads.mediation.MediationRewardedAd
 import com.google.android.gms.ads.mediation.MediationRewardedAdCallback
+import com.google.android.gms.ads.mediation.rtb.RtbSignalData
+import com.google.android.gms.ads.mediation.rtb.SignalCallbacks
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -64,6 +76,12 @@ class FyberMediationAdapterTest {
       FyberMediationAdapter.KEY_SPOT_ID to AdapterTestKitConstants.TEST_PLACEMENT_ID,
     )
   private val mockInitializationCompleteCallback: InitializationCompleteCallback = mock()
+  private val mockBannerAdLoadCallback:
+    MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback> =
+    mock()
+  private val mockInterstitialAdLoadCallback:
+    MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback> =
+    mock()
   private val mockRewardedAdLoadCallback:
     MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback> =
     mock()
@@ -253,6 +271,84 @@ class FyberMediationAdapterTest {
 
   // endregion
 
+  // region Collect Signals
+  @Test
+  fun collectSignals_invokesOnSuccessWithGeneratedToken() {
+    mockStatic(BidTokenProvider::class.java).use {
+      val mockSignalData = mock<RtbSignalData>()
+      val mockSignalCallback = mock<SignalCallbacks>()
+      whenever(BidTokenProvider.getBidderToken()) doReturn TEST_BID_RESPONSE
+
+      adapter.collectSignals(mockSignalData, mockSignalCallback)
+
+      verify(mockSignalCallback).onSuccess(TEST_BID_RESPONSE)
+    }
+  }
+
+  @Test
+  fun collectSignals_withNullToken_invokesOnSuccessEmptyString() {
+    mockStatic(BidTokenProvider::class.java).use {
+      val mockSignalData = mock<RtbSignalData>()
+      val mockSignalCallback = mock<SignalCallbacks>()
+      whenever(BidTokenProvider.getBidderToken()) doReturn null
+
+      adapter.collectSignals(mockSignalData, mockSignalCallback)
+
+      verify(mockSignalCallback).onSuccess("")
+    }
+  }
+
+  // endregion
+
+  // region Banner Ad Load Tests
+  @Test
+  fun loadRtbBannerAd_invokesLoadAd() {
+    mockStatic(InneractiveAdSpotManager::class.java).use {
+      val bannerAdConfiguration =
+        createMediationBannerAdConfiguration(context = activity, bidResponse = TEST_BID_RESPONSE)
+      val mockAdSpot = mock<InneractiveAdSpot>()
+      val mockInneractiveAdSpotManager =
+        mock<InneractiveAdSpotManager> { on { createSpot() } doReturn mockAdSpot }
+      whenever(InneractiveAdSpotManager.get()) doReturn mockInneractiveAdSpotManager
+
+      adapter.loadRtbBannerAd(bannerAdConfiguration, mockBannerAdLoadCallback)
+
+      mockInneractiveAdManager.verify {
+        InneractiveAdManager.setMediationName(eq(FyberMediationAdapter.MEDIATOR_NAME))
+        InneractiveAdManager.setMediationVersion(eq(MobileAds.getVersion().toString()))
+      }
+      verify(mockAdSpot).loadAd(eq(TEST_BID_RESPONSE))
+    }
+  }
+
+  // endregion
+
+  // region Banner Ad Load Tests
+  @Test
+  fun loadRtbInterstitialAd_invokesLoadAd() {
+    mockStatic(InneractiveAdSpotManager::class.java).use {
+      val interstitialAdConfiguration =
+        createMediationInterstitialAdConfiguration(
+          context = activity,
+          bidResponse = TEST_BID_RESPONSE,
+        )
+      val mockAdSpot = mock<InneractiveAdSpot>()
+      val mockInneractiveAdSpotManager =
+        mock<InneractiveAdSpotManager> { on { createSpot() } doReturn mockAdSpot }
+      whenever(InneractiveAdSpotManager.get()) doReturn mockInneractiveAdSpotManager
+
+      adapter.loadRtbInterstitialAd(interstitialAdConfiguration, mockInterstitialAdLoadCallback)
+
+      mockInneractiveAdManager.verify {
+        InneractiveAdManager.setMediationName(eq(FyberMediationAdapter.MEDIATOR_NAME))
+        InneractiveAdManager.setMediationVersion(eq(MobileAds.getVersion().toString()))
+      }
+      verify(mockAdSpot).loadAd(eq(TEST_BID_RESPONSE))
+    }
+  }
+
+  // endregion
+
   // region Rewarded Ad Load Tests
   @Test
   fun loadRewardedAd_setsMediationParametersAndInitializesInneractiveAdManager() {
@@ -331,6 +427,26 @@ class FyberMediationAdapterTest {
         FyberMediationAdapter.ERROR_DOMAIN,
       )
     verify(mockRewardedAdLoadCallback).onFailure(argThat(AdErrorMatcher(expectedAdError)))
+  }
+
+  @Test
+  fun loadRtbRewardedAd_invokesLoadAd() {
+    mockStatic(InneractiveAdSpotManager::class.java).use {
+      val rewardedAdParameters =
+        createMediationRewardedAdConfiguration(context = activity, bidResponse = TEST_BID_RESPONSE)
+      val mockAdSpot = mock<InneractiveAdSpot>()
+      val mockInneractiveAdSpotManager =
+        mock<InneractiveAdSpotManager> { on { createSpot() } doReturn mockAdSpot }
+      whenever(InneractiveAdSpotManager.get()) doReturn mockInneractiveAdSpotManager
+
+      adapter.loadRtbRewardedAd(rewardedAdParameters, mockRewardedAdLoadCallback)
+
+      mockInneractiveAdManager.verify {
+        InneractiveAdManager.setMediationName(eq(FyberMediationAdapter.MEDIATOR_NAME))
+        InneractiveAdManager.setMediationVersion(eq(MobileAds.getVersion().toString()))
+      }
+      verify(mockAdSpot).loadAd(eq(TEST_BID_RESPONSE))
+    }
   }
 
   // endregion